correct CanMsg ID to uint32_t from uint16_t

+++ include/can2serial/can2serial.h
@@ -31,7 +31,7 @@ typedef struct CanMsg
 		len = 8;
 		type = 0x03; //stdFrame
 	}
-	uint16_t ID;
+    uint32_t ID;
 	uint8_t len;
 	uint8_t data[8];
 	uint8_t type;
